USACO Training - transform
This problem made it clear to me the significance of the quote "Any fool can write code that a computer can understand. Good programmers write code that humans can understand." Originally, I had a solution that most likely would have worked (though I never confirmed it by testing it). However, when I showed it to someone for review, he stated that it was very hard to understand. I then worked my way to simplifying my code and through that, it was made clear to me just how much more helpful easy-to-understand code can be. Even for me as I was writing, I found it much easier to think through problems and debug my code.