USACO Traning - dualpal
This one was even faster than the previous, largely due to the fact that I was able to almost completely reuse the functions from the palsquare. As I fixed and optimized some things in my code, I noted how much more comfortable I felt as opposed to even just a few days ago. Hopefully this kind of progress and speed will be present even when I start to solve harder problems.