USACO Training - crypt1
This one took about 5 hours total to complete. Only about 1 and a half hours of that was actually developing the algorithm and writing code. The rest was mostly debugging. Luckily, (or not, depending on how you look at it,) the all the problems that took me a long time to debug were coding problems and not algorithm problems. I say luckily because it means that I just need to be a lot more careful when writing code. Also, as I get more experience, I'll be less likely to make those kinds of mistakes and/or fix them. Still, 5 hours is far from my goal. I'll take this as a valuable learning experience to BE CAREFUL of what I'm writing and how it could potentially affect my code.