USACO Training - combo
I finished this program relatively quickly, but that is in part because it was a lot easier compared to some of the previous problems. I could have been a lot faster, but I was messy when thinking of my algorithm. Although I had the right idea, I tried to rush and ended up doubting myself a lot when I wrote the actual code and made lots of unnecessary changes when I debugged because I didn't really know what certain parts did. I should take more care to first completely plan out what I want to do.