USACO Training - skidesign
This problem took less than half an hour to do and was a nice refresher after the difficulty I faced in the last problem (wormholes). But even with the lower difficulty, I still had some problems, so I really do need to be more careful and COMPLETELY think things through.