USACO Training - ariprog
This was the first program where I ran into a time problem, and ended up taking me about an entire day to solve. At first, I tried to optimize everything so that I ran certain functions as few times as possible. But in the end, it became apparent that there really was not much I could do. To fix the time problem, I had to trade space for speed. This solution was not as obvious to me as I think it should have been and I think it's important that I am able to solve problems similar to this much quicker than I did.