USACO Training - numtri
This problem also took me a while to solve. I had a solution that worked, but I didn't account for all the possibilities the problem presented. As a result, there were some test runs that led to some time problems. Finding a solution to the problem took some time, but I think as I grow more accustomed to fixing these issues, I will be able to get through them faster and faster.