USACO Training - pprime
This one took me a little bit of time to do. Most of this time was not spent coding the solution, but finding it. When reviewing the problem, there were many obvious but severe problems. The biggest one was time. Finding the solution to this problem could have actually been relatively easy, but I got myself stuck on a point because I went into solving the problem with some incorrect assumptions. From there it was relatively smooth sailing to completing the problem. From this I've learned I need to recognize what assumptions I'm making and whether or not those assumptions are absolutely correct.