USACO Training - sprime
I got to recycle a lot of my code from pprime with this, but it was still an interesting exercise. I ran into some problems because I didn't competely think out my algorithm. Although I think it unrealistic to expect to completely negate that, as it can be difficult to think of all possible solutions and potential outliers, I definitely do it too much right now and I hope I can begin to cut down on these kinds of mistakes.