USACO Training - castle
This castle problem is the first graphing problem in the USACO Training site. I had some problems thinking of how to solve this, though I'm not sure if it was because it's a new concept to me or if I just wasn't thinking properly. It is properly some combination of the two. After actually knowing what my solution would be, I was able to get the basics of it coded without much trouble. The details took a little more effort because I was careless in some areas, but overall it wasn't too difficult.