USACO Training - hamming
This problem was pretty tough for me. Thinking up a solution was difficult. I had many ideas, but some of them were not very efficient and sometimes just straight up didn't work. I think it'll be important for me to train myself to not think in circles so much, as that is one of the reasons it was so difficult for me.