USACO Training - preface
I used a really awkward solution to solve this one because I initially thought it would be simpler and easier to code. My thought process was that rather than generating each number, converting it to roman numerals, and then counting each roman numeral, I could instead instantly generate the amount of roman numerals using the fact that patterns exist.
In decimal writing, where after every 10 numbers, all digits - 0, 1, 2, 3, 4, 5, 6, 7, 8, and 9 - must have appeared. Similarly, after a certain amount of numbers, the roman numerals (given for the problem) - I, V, X, L, C, D, and M - must appear a certain number of times. I mistakenly thought it would be rather easy to determine and code the algorithm for the pattern.
While I cannot say if it took so long because of my inexperience in coding, the complexity of this solution, or simply my messy thought process, what I can say is that the time I needed to complete this problem was much more than it should've been. Thinking through the pattern was not very tricky but when it came to actually writing it down it started to become unclear. Multiple times I had to competely restart my thought process on the patterns because I doubted myself. It's likely I would've saved a lot of time if I had simply written it all down. All of this back and forth thinking not only slowed me down but even led to me completely forgetting to include certain parts of the algorithm, which then took me another full hour to completely debug.
From now on, I will try to always choose the simplest (but still efficient) solution I can. And if for some reason it is still complex, I've learned it is important to keep track of thoughts, so as not to waste time rethinking and reconfirming what doesn't need to be rethought or reconfirmed.