USACO Training - lamps
(lamps.in file committed with runround)
This problem was not too hard to think out but actually implementing everything was a little difficult for me. I think I was a bit tired, because although I clearly knew what I wanted to do, I kept going in circles with the details and ended up wasting some time fixing my initial mistakes.