USACO Training - money
This problem wasn't that hard, or at least shouldn't have been for me. It ended up taking me long time to debug because of some minor mistakes I made in both comprehension and just coding. However, I'm still pretty happy with the result as this was the first Dynamic Programming problem I was able to complete with (relative) ease.