USACO Training - maze1
Though this problem was fairly simple, I did run into some problems. One of them was coding the actual program. It wasn't that there were any difficulties in the algorithm, but rather that I confused myself by thinking in terms of x and y instead of rows and columns. It ended up in me reversing my arrays, which I had to fix. Also, I spent a decent amount of time fixing a very small mistake, and even while debugging, I initially overlooked it even though it was fairly obvious if given a bit of thought. This again comes back to my habit of not completely thinking things through.