USACO Training - cowtour
I spent too long on this problem. It wasn't any problem with coding or comprehending what the problem was about, but rather a fundamental problem with how I approached and attempted to solve it. In my thinking, I heavily doubt that I completely missed the mistake, but I definitely paid little attention to it, which costed me a lot of itme. In the end, it again had to do with an assumption that I had falsely verified to myself many many times. I don't really undestand why or how I didn't realize it sooner now that I look back and see how simple it was. But this has been a great lesson for me overall.