USACO Training - inflate
I struggled with this problem a fair bit. I used Dynamic Programming to solve this, but I had a lot of trouble. Part of it was because of the little break I had taken, and I had to remind myself of some key concepts, but most of my struggle came from me simply not being able to think of a good solution. However, I decided to write my code anyway, and to my surprise, it worked surprisingly well. From then on, I decided it was good enough, and spent the rest of my time optimizing my code so it would fit the time requirements.
It worked, but I realize it was not the optimal solution. I hope that as I gain more experience I'll be able to deal with problems like this more easily.