USACO Training - contact
This problem was actually quite simple. Writing it was a little bit of a pain, but I suspect it was because I went about it in a weird way. Apart from that, this problem shouldn't have taken me very long, but I had to check it multiple times and debug it because I didn't pay attention to the problem. Luckily, it wasn't anything about the content but rather how it was supposed to be outputted, but this shouldn't be happening at all. I had to go back to the problem page to check the output requirements and almost every tme I did I found I had missed something before and had to edit my code for it. I'll have to be a lot more attentive in the future so this doesn't happen.