USACO Training - stamps
When first trying to solve this problem, I realized it was DP but didn't actually think of what that meant. My initial solution didn't quite work, and a result, I had to competely rethink my solution. After that, it was simply a matter of dealing with some issues on the site where I turned this in. That took some time too but it was more of a minor annoyance than any real challenge.