USACO Training - fact4
This problem was not hard, but the way I solve it was not very good. Instead of dealing with 5's and 2's, which were giving me problems, I just decided to modulo everything by a very large number. I should think through my problems and identify why exactly it is a problem and what is causing it. If I did that here, I think I would've come to the conclusion that 5's and 2's were the only real issue. From there, it wouldn't have been hard to think up a solution for that.