USACO Training - spin
This problem was not very hard. The most difficult thing about this problem was dealing with the fact that integers are not the best at representing points on a circle because circles loop. Even then, that was a fairly simple fix. From there, the biggest road bump was simply that I had misinterpreted a part of the problem. While fixing that I didn't pay close enough attention to my edits, which then led to another few minutes of debugging. I don't think I've made these small mistakes as often as before, but I hope to even still continue to reduce the amount of times this happens.