USACO Training - ratios
This problem was very simple, or at least it seemed that way to me at first. And the solution is still relatively simple, just not as much as it had first appeared to me. It didn't take me too long to write most of it, but there was one edge condition that I hadn't accounted for. Specifically, when one of the feeds equaled zero. I used a lot of division and when something equaled zero, it simply wouldn't work because it would try to divide by it. I didn't have to spend a lot of time accounting for this edge condition, but I realize it's important that in the future I am able to think of and prepare for these kinds of situations beforehand.