USACO Training - msquare
At first I struggled thinking of how to do this. For some reason I didn't believe breadth-first search would work and shut it out from my mind completely. It was only upon coming back much later that I realized it was a very viable solution. After that, coding it was tedious because I was not too clear about how exactly I was going to solve the problem, but my idea was clear enough that there were no real difficulties other than figuring some details out along the way. The only issue was me realizing that there was a memory problem because the site I turned it into had a limited memory, but that was pretty easy to find a solution to.