USACO Training - butter
This problem took me a while. I was able to think of potential solutions and ended up writing one of them, but I couldn't quite get past the time problem. In the end, I had to look for a new method, which was using Dijkstra's Algorithm but using a heap instead of a table, which I had not done before. Doing that was able to cut the time down just barely enough, so I was delighted to see in the analysis that this was, in fact, the intended solution.