USACO Training - fence
This problem was quite simple. I used the eurelian tours pseudocode given in the text as a template and it worked beautifully. It did require some fine tuning because of some minor inconvenience (such as it storing the answer backwards) but that really was no problem at all. I think it was really nice and helped me get a little familiar with the Eurelian Paths algorithm.