USACO Training - shopping
This problem was pretty different from others and my solution was a little weird, both in the concept of it and the actual execution. My solution did work, but I had to take a look at the analysis. Surprisingly, one solution was to turn it into a shortest paths problem. It had never occurred to me that this problem could be anything like that. I think it'll be important for me to be able to think of that in the future, as that could've saved me some time and effort.