USACO Training - camelot
The solution to this problem was actually surprisingly simple. It took me a fairly long time to get there but I managed to do it after multiple tries. A lot of the time spent for this problem actually came from details and me not being careful when writing my code. Hopefully from now on I'll be more accepting towards the very simple solutions, even if at first glance they do not seem very viable. Also, I'll need to be more careful about writing my code. Mistakes (especially small and barely noticeable ones such as swapping the two pointers when accessing a 2d array) can have a huge impact on the program and be hard to find within multiple lines of code.