USACO Training - heritage
This problem was pretty simple. It did take me a while to figure out the relationship between the three orders of traversal, but once I cleared that up thinking of a way to get to the solution was not very hard. The hardest part was actually writing it, which I've found is pretty common for me. I think I should work on thinking through solutions not just from a logical standpoint but also from a coding standpoint so I do not spend so much time converting my logic into actual code.