USACO Training - fence6
This problem seemed quite intimidating at first, but breaking it down, I found it was actually not very complex at all. However, I will admit my code is a little messy. The code actually follows my thinking very closely. But in this case, that was probably not the best idea neatness-wise. For example, converting the input from fences to endpoints was probably not needed in the code, but it helped me a lot in thinking things through, so it was added in anyway.