USACO Training - ditch
This problem shouldn't have been hard, but because of a small error on my part, it took me many hours. The error was in me misunderstanding part of the problem which caused my input to be a tiny bit off. It wasn't until I deceided to go through the fairly large input that I was able to idenitfy what the mistake was.