USACO Training - job
The solution for the problem was pretty "out there" for me, as in it would've never been a first thought when thinking of a potential solution. However, after stumbling upon it, it seemed rather abstract. I then spent a bit of time to fully understand how and why exactly it worked. I think this was a nice little lesson to teach me to be more flexible in how I approach problems.