USACO Training - race3
This problem was not very difficult, but it did require some thinking. I was able to finish a piece of code that worked for many test cases, but when met with the lengthier onces, proved too slow. Way too slow. It took me a lot of time to think of the relatively solution. I think I can attribute this to a lack of experience for now, but I realize this must change in the future.