USACO Training - shuttle
This problem was pretty confusing at first, because I tried using algorithms to solve it. In the end, I essentially gave up and tried what I considered to be a cheesy approach, which ended up being the intended solution. I realized that I needed to broaden my view when it came to looking for solutions.