USACO Training - snail
This problem was more straightforward that I initially expected. Looking at the problem, a simple solution is not hard to think of. However, I had thought that brute force would not work and would be too slow. For a good amount of time, I tried to think of some other solution that would work better, but I could not think of one at all. I eventually tried the simplest solution, and it ended up working. It felt a little like a cold bucket of water in the face to me, and I saw that I had a habit of almost completely disregarding straightforward solutions.