USACO Training - schlnet
This problem seemed really complex to me at first. It took me a while to work out a solution, and when I did, writing also wasn't the easiest, although it wasn't too hard either. The solution ended up working, but after reading the analysis, I felt a little silly. The solution on the website is essentially mine, but it was different in a way. Upon closer inspection and some thinking, I realized that the part that had given me the most trouble in my solution was actually completely unnecessary, and the solution would have been the same without it.
It wasn't all that bad, as I was still able to gain some experience writing the algorithm I used, but I was still a little shocked I hadn't realized how over-the-top my solution was sooner. I think this caan be seen as me overcomplicating things again.