USACO Training - tour
My main issue here was not actually reading the problem. Initially, I'm not sure if I was tired or simply not paying much attention, but I completely missed a pretty big part of the problem, and worked with my incomplete view for far too long. I'm going to need to be far more attentive.