USACO Training - charrec
The problem initially was pretty scary to me. I can't exactly pinpoint what exactly made me feel this way, but something about the problem made me fee like it was going to be really tough to solve. But after thoroughly thinking it through, I came to the realization that it's just a simple DP problem. I realized there's a lot of value in getting to reaallly know a problem, as it was only after I had completely solidified my understanding of it that I was able to see how simple it really was.