USACO Training - picture
My initial solution to this was quite, in a word, convoluted. It ended up being too slow anyway, but even after changing it to a solution that worked, it still wasn't too simple. Still, I was pretty happy with it until I read the analysis.
After reading the analysis, I was in awe. While I still considered my solution to be about as conceptually similar (about as easy to understand), their solution was much easier to read and many times more efficient.