USACO Training - twofive
My initial solution was pretty simple with a tiny optimization. The optimization didn't do much, and I ended up having to completely ditch that solution. Getting to my final solution involved a way of thinking that I found I wasn't used to. It wasn't actually very complex, but it was not as obvious to me as I think it should have been. It served as another reminder that I need to learn to broaden my view when thinking about problems.